pyproject.toml
[project]
name = "patch"
version = "0.1.0"
description = "Collect GitHub issue/PR diffs into a code SFT dataset"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "aiofiles>=23.0",
    "datasets>=2.18",
    "huggingface-hub>=0.22",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "orjson>=3.9",
    "tenacity>=8.2",
]

[project.scripts]
collect = "scripts.run_collect:main"
process = "scripts.run_process:main"
merge = "scripts.run_merge:main"
push = "scripts.run_push:main"
peek = "scripts.run_peek:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/patch", "scripts", "config"]
